backend/requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.7.0
docx2txt==0.8
Flask==3.0.3
Flask-Cors==4.0.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
html5lib==1.1
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
Markdown==3.6
markdown2==2.4.13
MarkupSafe==2.1.5
oscrypto==1.3.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pyparsing==3.1.2
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
rsa==4.9
six==1.16.0
svglib==1.5.1
tinycss2==1.3.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
xhtml2pdf==0.2.16
